fix(build): mac-clang preset fails to build on Apple Clang 17
Cause

libc++ 19.1 gates <stop_token> behind _LIBCPP_ENABLE_EXPERIMENTAL. The
header declares nothing, so std::stop_token fails to resolve in
st::devices::ets. libstdc++ and the MSVC STL ship <stop_token> ungated,
so this does not affect other presets.

Three warnings are fatal under -Werror with Apple Clang:

- -Wsign-conversion in splice_sh2a.cpp and test_native_properties.cpp
- -Wignored-attributes on the gnu_printf archetype in widgets.hpp

Fix

- CMakeLists.txt probes for std::stop_token and defines
  _LIBCPP_ENABLE_EXPERIMENTAL when the probe needs it. Probing the type,
  not the compiler id, leaves the define off standard libraries that
  ship <stop_token> ungated. Configure fails when neither probe compiles.
- The two shifts cast to uint32_t before the OR.
- widgets.hpp uses the printf archetype under Clang and gnu_printf under
  GCC.

Validation

ctest --preset mac-clang passes except the two CLI audit stats cases,
which read fixtures/demo.stune/audit.log. .gitignore excludes that file,
so they fail on a fresh clone of main.

+++ src/feature_patch/src/splice_sh2a.cpp
@@ -36,7 +36,7 @@ namespace {
 
 // JMP @Rn — 0100 nnnn 0010 1011 = 0x402B | (n << 8). For R0: 0x402B.
 [[nodiscard]] constexpr std::uint16_t enc_jmp_at_reg(std::uint8_t reg) noexcept {
-    return static_cast<std::uint16_t>(0x402BU | (static_cast<std::uint16_t>(reg) << 8U));
+    return static_cast<std::uint16_t>(0x402BU | (static_cast<std::uint32_t>(reg) << 8U));
 }
 
 // SH-2A is big-endian on the wire. Write a u16 in BE order.
